refactor(notifications): move alert_firing dispatch from Celery to Temporal
